+++ Team41/Code41/src/unit_testing/src/SimpleParser/TestTokenizerUtils.cpp
@@ -11,7 +11,7 @@ void TestTokenizerUtils::tokenizeAndCompare(string s, vector<pair<TokenType, str
     Tokenizer tokenizer(move(s));
     Tokens tokens = tokenizer.tokenize();
     REQUIRE(tokens.size() == expected.size());
-    for (int i = 0; i < tokens.size(); i++) {
+    for (size_t i = 0; i < tokens.size(); i++) {
         auto [expectedType, expectedVal] = expected[i];
         Token t = tokens[i];
         TokenType gotType = t.getType();
